Create data files (.dat and .pin) for use with multispecies model, Hydra.

Initial data files have been lazily loaded. To view their contents please run the create_documentation rmd file

create_datpin_files(listOfParameters, dataList)

Arguments

listOfParameters

see setup_default_inputs

dataList

A data set comprising all necessary inputs. This data set can be the lazily loaded data set or it can be modified in any way to represent a faux assessment. See hydraDataList for details

Value

updated list of lazily loaded data "hydraData"

Output

Two data files are output. A ".dat" and a ".pin" file. These are required by the Hydra code to run. The name of the file is determined by the variable outputFilename in the list listOfParameters. By default the files are saved in the current working directory. You can specify a differnt location in listofParameters. See setup_default_inputs