sets up Harvest Control Rules (HCR)

Harvest control rules are evaluated using either a step function or a continuous linear function. For a continuous function all that is required is minMaxThreshold and minMaxExploitation. For a step function, Nthresholds, thresholds, exploitationOptions are also required. This function with create the step function thresholds and exploitationOptions given minMaxThreshold, minMaxThreshold and Nthresholds

set_hcr(
  currentData = hydradata::hydraDataList,
  Nthresholds = 5,
  minMaxThresholds = c(0.1, 0.4),
  minMaxExploitations = c(0.05, 0.4),
  increment = 0.05
)

Arguments

currentData

List: The current hydraData in which changes need to be made

Nthresholds

Numeric scalar: Number of thresholds in the step function (Default = 5)

minMaxThresholds

Numericric vector (1x2): Specifying the minimum and maximum threshold levels (proportion of B0) for management

minMaxExploitations

Numeric vector (1x2): Specifying the minimum and maximum exploitation levels at minMaxThresholds

increment

Numeric Scalar: increment to evaluate the number of scenarios between minMaxExploitations

Value

A list identical to hydraDataList with the fields below updated

Updated fields

exploitationOptions - Numeric matrix: (Nthresholds x NScenarios). Range of possible scenarios for harvest control rule

thresholds - Numeric vector (1x Nthresholds): Specifying the threshold levels for which a change in exploitation will occur

Nthresholds - Numeric scalar: Specifying the number of thesholds

minMaxThresholds - Numeric vector (1x2): Specifying the minimum and maximum threshold levels (proportion of B0) for management

Usage

Run this function to change the default set of step function scenarios.